# packages: misc postgres performance + correctness improvements