--- a/internal/codeintel/dependencies/internal/store/store_test.go
+++ b/internal/codeintel/dependencies/internal/store/store_test.go
@@ -114,6 +114,10 @@ func TestListPackageRepoRefs(t *testing.T) {
 			{Scheme: "npm", Name: "applesauce", Versions: []string{"1.2.3"}},
 			{Scheme: "somethingelse", Name: "banana", Versions: []string{"0.1.2"}},
 		},
+		// should not be listed due to no versions
+		{
+			{Scheme: "npm", Name: "burger", Versions: []string{}},
+		},
 	}
 
 	for _, insertBatch := range batches {
@@ -124,8 +128,16 @@ func TestListPackageRepoRefs(t *testing.T) {
 
 	var lastID int
 	for i, test := range [][]shared.PackageRepoReference{
-		{{Scheme: "npm", Name: "bar"}, {Scheme: "npm", Name: "foo"}, {Scheme: "npm", Name: "banana"}},
-		{{Scheme: "npm", Name: "turtle"}, {Scheme: "npm", Name: "applesauce"}, {Scheme: "somethingelse", Name: "banana"}},
+		{
+			{Scheme: "npm", Name: "bar", Versions: []shared.PackageRepoRefVersion{{Version: "2.0.0"}, {Version: "2.0.1"}, {Version: "3.0.0"}}},
+			{Scheme: "npm", Name: "foo", Versions: []shared.PackageRepoRefVersion{{Version: "1.0.0"}}},
+			{Scheme: "npm", Name: "banana", Versions: []shared.PackageRepoRefVersion{{Version: "2.0.0"}}},
+		},
+		{
+			{Scheme: "npm", Name: "turtle", Versions: []shared.PackageRepoRefVersion{{Version: "4.2.0"}}},
+			{Scheme: "npm", Name: "applesauce", Versions: []shared.PackageRepoRefVersion{{Version: "1.2.3"}}},
+			{Scheme: "somethingelse", Name: "banana", Versions: []shared.PackageRepoRefVersion{{Version: "0.1.2"}}},
+		},
 	} {
 		depRepos, total, hasMore, err := store.ListPackageRepoRefs(ctx, ListDependencyReposOpts{
 			Scheme: "",
@@ -152,7 +164,11 @@ func TestListPackageRepoRefs(t *testing.T) {
 
 		for i := range depRepos {
 			depRepos[i].ID = 0
-			depRepos[i].Versions = nil
+			for j, version := range depRepos[i].Versions {
+				depRepos[i].Versions[j] = shared.PackageRepoRefVersion{
+					Version: version.Version,
+				}
+			}
 		}
 
 		if diff := cmp.Diff(test, depRepos); diff != "" {