--- a/migrations/frontend/1678409821_package_repos_missing_indexes/metadata.yaml
+++ b/migrations/frontend/1678409821_package_repos_missing_indexes/metadata.yaml
@@ -1,0 +1,2 @@
+name: package_repos_missing_indexes
+parents: [1678220614, 1678320579, 1678380933]