Changes include:
1. Reintroducing the unique index `(scheme,name)` that was originally `(scheme,name,version)` but was dropped when the `version` column was dropped. Not problematic as our insert query guaranteed no duplicates anyways, but nice-to-have
2. Improved query for listing package repo refs. The changes introduced [here](https://github.com/sourcegraph/sourcegraph/pull/48037/files#diff-6dff69941489dc9abde8669bfeaa09319bc712b1745e54c17da8f051e663ca4fR93-R131) caused performance to drop dramatically (see Grafana Screenshot below, after 03/09 00:00 less concentration in the 4th row from the bottom, instead spreading out further to higher rows where higher = slower)
3. Added indexes to `(scheme,id ASC)` and `(name,id ASC)`. Having `id` in these columns helps due to use always having `ORDER BY id`, skipping potential explicit sort steps, as well as the above in due to frequent filtering by scheme or name. Unique index above wouldnt be used due to the postgres preferring to use the primary key index for its sorted `id` column property and filtering on scheme/name instead.

See New Plan and Old Plan below for details.

<details>
<summary>Grafana Screenshot</summary>

![image](https://user-images.githubusercontent.com/18282288/224200497-d3c594a7-50f2-4a52-87cd-a8b33fc5a51e.png)

</details>

<details>
<summary>New Plan</summary>

[Visualize on explain.dalibo.com](https://explain.dalibo.com/plan/4c3g61g7hea992fh)

```sql
EXPLAIN (ANALYZE, BUFFERS, VERBOSE) SELECT
    lr.id,
    lr.scheme,
    lr.name,
    lr.blocked,
    lr.last_checked_at,
    array_agg(prv.id ORDER BY prv.id) as vid,
    array_agg(prv.version ORDER BY prv.id) as version,
    array_agg(prv.blocked ORDER BY prv.id) as vers_blocked,
    array_agg(prv.last_checked_at ORDER BY prv.id) as vers_last_checked_at
FROM lsif_dependency_repos lr
JOIN LATERAL (
    SELECT id, package_id, version, blocked, last_checked_at
    FROM package_repo_versions
    WHERE package_id = lr.id
    ORDER BY id
) prv
ON lr.id = prv.package_id
WHERE scheme = 'rust-analyzer'
GROUP BY lr.id
ORDER BY lr.id
LIMIT 500;
```

```
Limit  (cost=52.02..5985487.30 rows=18 width=170) (actual time=0.083..44.818 rows=500 loops=1)
"  Output: lr.id, lr.scheme, lr.name, lr.blocked, lr.last_checked_at, (array_agg(prv.id ORDER BY prv.id)), (array_agg(prv.version ORDER BY prv.id)), (array_agg(prv.blocked ORDER BY prv.id)), (array_agg(prv.last_checked_at ORDER BY prv.id))"
  Buffers: shared hit=33172
  ->  GroupAggregate  (cost=52.02..5985487.30 rows=18 width=170) (actual time=0.082..44.739 rows=500 loops=1)
"        Output: lr.id, lr.scheme, lr.name, lr.blocked, lr.last_checked_at, array_agg(prv.id ORDER BY prv.id), array_agg(prv.version ORDER BY prv.id), array_agg(prv.blocked ORDER BY prv.id), array_agg(prv.last_checked_at ORDER BY prv.id)"
        Group Key: lr.id
        Buffers: shared hit=33172
        ->  Nested Loop  (cost=52.02..5985486.72 rows=18 width=66) (actual time=0.042..26.712 rows=14969 loops=1)
"              Output: lr.id, lr.scheme, lr.name, lr.blocked, lr.last_checked_at, prv.id, prv.version, prv.blocked, prv.last_checked_at"
              Buffers: shared hit=33172
              ->  Index Scan using noah_test_lsif_dependency_repos_scheme_id on public.lsif_dependency_repos lr  (cost=0.42..27979.76 rows=114057 width=42) (actual time=0.020..0.718 rows=501 loops=1)
"                    Output: lr.id, lr.scheme, lr.name, lr.blocked, lr.last_checked_at"
                    Index Cond: (lr.scheme = 'rust-analyzer'::text)
                    Buffers: shared hit=1004
              ->  Subquery Scan on prv  (cost=51.59..52.22 rows=1 width=32) (actual time=0.034..0.042 rows=30 loops=501)
"                    Output: prv.id, prv.package_id, prv.version, prv.blocked, prv.last_checked_at"
                    Filter: (lr.id = prv.package_id)
                    Buffers: shared hit=32168
                    ->  Sort  (cost=51.59..51.70 rows=42 width=32) (actual time=0.033..0.036 rows=30 loops=501)
"                          Output: package_repo_versions.id, package_repo_versions.package_id, package_repo_versions.version, package_repo_versions.blocked, package_repo_versions.last_checked_at"
                          Sort Key: package_repo_versions.id
                          Sort Method: quicksort  Memory: 25kB
                          Buffers: shared hit=32168
                          ->  Index Scan using package_repo_versions_unique_version_per_package on public.package_repo_versions  (cost=0.43..50.46 rows=42 width=32) (actual time=0.003..0.025 rows=30 loops=501)
"                                Output: package_repo_versions.id, package_repo_versions.package_id, package_repo_versions.version, package_repo_versions.blocked, package_repo_versions.last_checked_at"
                                Index Cond: (package_repo_versions.package_id = lr.id)
                                Buffers: shared hit=32168
Planning Time: 0.227 ms
Execution Time: 44.908 ms

```

</details>

<details>
<summary>Old Plan</summary>

[Visualize on explain.dalibo.com](https://explain.dalibo.com/plan/d2709418e14e89e7)

```sql
EXPLAIN (ANALYZE, VERBOSE, BUFFERS) SELECT
      lr.id,
      lr.scheme,
      lr.name,
      lr.blocked,
      lr.last_checked_at,
      array_agg(prv.id ORDER BY prv.id) as vid,
      array_agg(prv.version ORDER BY prv.id) as version,
      array_agg(prv.blocked ORDER BY prv.id) as vers_blocked,
      array_agg(prv.last_checked_at ORDER BY prv.id) as vers_last_checked_at
FROM lsif_dependency_repos lr
JOIN package_repo_versions prv
ON lr.id = prv.package_id
WHERE scheme = 'rust-analyzer' AND TRUE
GROUP BY lr.id, lr.scheme, lr.name
ORDER BY lr.id ASC
LIMIT 500;
```

```
Limit  (cost=0.85..1093.56 rows=500 width=170) (actual time=294.329..413.149 rows=500 loops=1)
"  Output: lr.id, lr.scheme, lr.name, lr.blocked, lr.last_checked_at, (array_agg(prv.id ORDER BY prv.id)), (array_agg(prv.version ORDER BY prv.id)), (array_agg(prv.blocked ORDER BY prv.id)), (array_agg(prv.last_checked_at ORDER BY prv.id))"
  Buffers: shared hit=903886
  ->  GroupAggregate  (cost=0.85..249784.56 rows=114295 width=170) (actual time=294.328..413.072 rows=500 loops=1)
"        Output: lr.id, lr.scheme, lr.name, lr.blocked, lr.last_checked_at, array_agg(prv.id ORDER BY prv.id), array_agg(prv.version ORDER BY prv.id), array_agg(prv.blocked ORDER BY prv.id), array_agg(prv.last_checked_at ORDER BY prv.id)"
        Group Key: lr.id
        Buffers: shared hit=903886
        ->  Merge Join  (cost=0.85..240311.37 rows=574983 width=66) (actual time=294.270..392.733 rows=14969 loops=1)
"              Output: lr.id, lr.scheme, lr.name, lr.blocked, lr.last_checked_at, prv.id, prv.version, prv.blocked, prv.last_checked_at"
              Merge Cond: (lr.id = prv.package_id)
              Buffers: shared hit=903886
              ->  Index Scan using noah_test_lsif_dependency_repos_scheme_id on public.lsif_dependency_repos lr  (cost=0.42..27826.97 rows=114295 width=42) (actual time=0.019..0.360 rows=501 loops=1)
"                    Output: lr.id, lr.scheme, lr.name, lr.blocked, lr.last_checked_at"
                    Index Cond: (lr.scheme = 'rust-analyzer'::text)
                    Buffers: shared hit=226
              ->  Index Scan using package_repo_versions_unique_version_per_package on public.package_repo_versions prv  (cost=0.43..203194.13 rows=1329133 width=32) (actual time=0.004..334.291 rows=484029 loops=1)
"                    Output: prv.id, prv.version, prv.blocked, prv.last_checked_at, prv.package_id"
                    Buffers: shared hit=903660
Planning Time: 0.382 ms
Execution Time: 413.239 ms
```

</details>

## Test plan

Lots of querying on dotcom data, back n forth with the Eri{c,k}s, additional unit test coverage